Use ConvertToLLVM2 for both linalg tensor and buffer passes (#4681)
