+++ iree/compiler/Conversion/LinalgToLLVM/CMakeLists.txt
@@ -23,7 +23,6 @@ iree_cc_library(
   SRCS
     "ConvImg2ColMatmulConversion.cpp"
     "ConvertToLLVM.cpp"
-    "ConvertToLLVM2.cpp"
     "KernelDispatch.cpp"
     "LinalgBufferizePass.cpp"
     "LinalgRewriteDestructiveUpdatesPass.cpp"
